requirements.txt
beautifulsoup4==4.12.2
common==0.1.2
numpy==1.22.4
Pillow==10.0.0
requests==2.27.1
scikit_learn==1.3.0
six==1.16.0
tensorflow==2.13.0
tqdm==4.64.0
transformers==4.28.1
